Add storage of trades and related tables to C*.
All trades are stored into wide rows which grow up to a fixed limit,
allowing efficient fetching of trades given transaction IDs.

In addition to storing trades there are also tables for managing
the latest retrieved content and such.

For example it is possible to retrieve the transaction IDs associated
with a given market on some day.